Separate tuple params deserializer into its own module and add a test
