+++ ethabi/src/event_param.rs
@@ -1,10 +1,10 @@
 //! Event param specification.
 
-use serde::de::{Error, MapAccess, SeqAccess, Visitor};
+use serde::de::{Error, MapAccess, Visitor};
 use serde::{Deserialize, Deserializer};
-use serde_json::Value;
 use std::{fmt};
 use ParamType;
+use TupleParams;
 
 /// Event param specification.
 #[derive(Debug, Clone, PartialEq)]
@@ -15,19 +15,6 @@ pub struct EventParam {
 	pub kind: ParamType,
 	/// Indexed flag. If true, param is used to build block bloom.
 	pub indexed: bool,
-}
-
-pub struct TupleParams {
-    params: Vec<Box<ParamType>>,
-}
-
-impl<'a> Deserialize<'a> for TupleParams {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'a>,
-    {
-        deserializer.deserialize_seq(TupleParamsVisitor)
-    }
 }
 
 impl<'a> Deserialize<'a> for EventParam {
@@ -94,7 +81,7 @@ impl<'a> Visitor<'a> for EventParamVisitor {
                 if let ParamType::Tuple(_) = param_type {
                     let tuple_params= components
                         .ok_or_else(|| Error::missing_field("components"))?;
-                    Ok(ParamType::Tuple(tuple_params.params))
+                    Ok(ParamType::Tuple(tuple_params.0))
                 } else {
                     Ok(param_type)
                 }
@@ -105,31 +92,6 @@ impl<'a> Visitor<'a> for EventParamVisitor {
             kind,
             indexed,
         })
-    }
-}
-
-struct TupleParamsVisitor;
-impl<'a> Visitor<'a> for TupleParamsVisitor {
-    type Value = TupleParams;
-
-    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        write!(formatter, "a valid event parameter spec")
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: SeqAccess<'a>,
-    {
-        let mut params: Vec<Box<ParamType>> = Vec::new();
-
-        while let Some(param) = seq.next_element()? {
-            let p: Value = param;
-            let kind: &Value = p.get("type")
-                .ok_or_else(|| Error::custom("Invalid tuple param type"))?;
-            params.push(Box::new(ParamType::deserialize(kind).unwrap()));
-        }
-
-        Ok(TupleParams { params })
     }
 }
