+++ ethabi/src/lib.rs
@@ -34,6 +34,7 @@ mod log;
 mod operation;
 mod param;
 mod signature;
+mod tuple_params;
 mod util;
 
 pub use param_type::ParamType;
@@ -49,6 +50,7 @@ pub use param::Param;
 pub use log::{Log, RawLog, LogParam, ParseLog, LogFilter};
 pub use event::Event;
 pub use event_param::EventParam;
+pub use tuple_params::TupleParams;
 
 /// ABI address.
 pub type Address = ethereum_types::Address;
